authorization of jwt

+++ controllers/api/v1/posts_api.js
@@ -25,18 +25,19 @@ module.exports.destroy = async function(req,res){
     try{
         let post = await Post.findById(req.params.id);
         //.id is used instead of ._id so that it can be converted into string
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
             await Comment.deleteMany({post : req.params.id});
             
             res.json(200,{
                 message : 'Post and associated comments deleted successfully'
             });
-        // }
-        // else{
-        //     req.flash('error','Not authorized to delete this post');
-        //     return res.redirect('back');
-        // }
+        }
+        else{
+            return res.json(401,{
+                message : "Unauthorized"
+            })
+        }
     }catch(err){
         console.log('*******',err);
         res.json(500,{
